+++ wgpu-hal/src/vulkan/mod.rs
@@ -31,11 +31,13 @@ mod descriptor;
 mod device;
 mod drm;
 mod instance;
+mod pnext_chain;
 mod sampler;
 mod semaphore_list;
 mod swapchain;
 
 pub use adapter::PhysicalDeviceFeatures;
+pub(crate) use pnext_chain::PnextChain;
 
 use alloc::{boxed::Box, ffi::CString, sync::Arc, vec::Vec};
 use core::{
@@ -356,40 +358,6 @@ impl Surface {
             .downcast_ref::<swapchain::NativeSurface>()
             .expect("Surface should be a native Vulkan surface");
         unsafe { surface.set_next_swapchain_create_chain(chain) };
-    }
-}
-
-/// A caller-provided `pNext` chain, stashed by one of the `set_next_*_chain`
-/// setters until the Vulkan call that consumes it.
-pub(crate) struct PnextChain(*mut vk::BaseOutStructure<'static>);
-
-// SAFETY: The pointer is only dereferenced at the Vulkan call that consumes the
-// chain. Each setter's contract keeps the chain valid and unaliased until then.
-unsafe impl Send for PnextChain {}
-unsafe impl Sync for PnextChain {}
-
-impl PnextChain {
-    /// Wraps the raw chain pointer that a `set_next_*_chain` setter received.
-    pub(crate) fn new(chain: *mut c_void) -> Self {
-        Self(chain.cast())
-    }
-
-    /// Splices this chain in front of `existing`, the current `p_next` of a
-    /// Vulkan info struct, and returns the new chain head.
-    ///
-    /// # Safety
-    ///
-    /// The chain must still be valid, and the info struct must be passed to the
-    /// Vulkan call that consumes the chain.
-    pub(crate) unsafe fn splice_into(self, existing: *const c_void) -> *const c_void {
-        unsafe {
-            let mut tail = self.0;
-            while !(*tail).p_next.is_null() {
-                tail = (*tail).p_next;
-            }
-            (*tail).p_next = existing.cast_mut().cast();
-        }
-        self.0.cast()
     }
 }
